Feat: udistrital/sga_documentacion#25 correción modelo
se corrige el modelo de respuesta y logica implementada
